feat(mobile): long-form, locale-aware dates for all date-value displays
Show date values as text in long form (the month spelled out — "19. Juni
2026" / "June 19, 2026") instead of the cryptic numeric short form, while
keeping the device/app locale. Covers every timestamp + date-value display:
last-synced (Sync, Contacts, Contacts settings), external-cache last-updated,
snapshot + sync-protocol timestamps, task completed-at, conflict timestamps,
reminder trigger times, search-result dates, task due/scheduled/resurface
dates, the deadline/scheduled chips on calendar days, and the Week/Agenda
range headers.

Also fixes a latent locale bug: the four Sync-screen timestamps called
toLocaleString() with NO locale (system default, ignoring the app's language)
— they now go through the shared formatLongDateTime(date, i18n.language)
helper. The calendar full-date headers already used a long month and the
weekday/month name pickers are untouched (they're not date values).

+++ mobile/src/screens/AgendaScreen.tsx
@@ -118,7 +118,7 @@ export default function AgendaScreen({
   );
   const fmtShortDate = useCallback(
     (d: Date) =>
-      d.toLocaleDateString(i18n.language, { year: 'numeric', month: 'short', day: 'numeric' }),
+      d.toLocaleDateString(i18n.language, { year: 'numeric', month: 'long', day: 'numeric' }),
     [i18n.language],
   );
   const timeLabel = useCallback(
