+++ mobile/src/screens/WeekScreen.tsx
@@ -94,7 +94,7 @@ export default function WeekScreen({ navigation, route }: RootStackScreenProps<'
 
   const fmtShortDate = useCallback(
     (d: Date) =>
-      d.toLocaleDateString(i18n.language, { year: 'numeric', month: 'short', day: 'numeric' }),
+      d.toLocaleDateString(i18n.language, { year: 'numeric', month: 'long', day: 'numeric' }),
     [i18n.language],
   );
